Fix signedness conversion warnings in image library
- Add explicit (unsigned int) casts for DefaultDepth macro usage
  in libxli.c, xloadimage.c, root.c, window.c (3 places)
- Add explicit (unsigned int) cast for DefaultDepth in create.c (libXPM)
- Add explicit (unsigned int) cast for DefaultDepth in glib_x.c (viola)
- Fix signedness issues in value.c reverse loops by casting len to int
- Modernize flipBits() function from K&R to ANSI C style
- Fix incorrect variable type in doMemToValLSB (int->unsigned long)

DefaultDepth macro from X11 returns int, but functions expect unsigned int.
Since depth values are always non-negative, explicit casts are safe.
Reverse loop counters must remain signed int to properly detect negative
values and avoid infinite loops with unsigned underflow.

+++ src/libIMG/libxli.c
@@ -183,7 +183,7 @@ int loadImg(char* imagename, /*Image** image,*/ XImageInfo** ximageinfo, int* wi
         if (DefaultDepth(disp, scrn) == 1)
             dispimage = newBitImage(winwidth, winheight);
         else {
-            dispimage = newRGBImage(winwidth, winheight, DefaultDepth(disp, scrn));
+            dispimage = newRGBImage(winwidth, winheight, (unsigned int)DefaultDepth(disp, scrn));
             dispimage->rgb.used = 1;
         }
         *(dispimage->rgb.red) = 65535; /* default border value is white */
